requirements.txt
isodate==0.6.0
pyparsing==2.4.0
PyYAML==5.1.1
rdflib==4.2.2
six==1.12.0
